some bugfixes related to cleanup and taking averages for island weights again

+++ source/blender/blenloader/intern/writefile.c
@@ -1402,7 +1402,7 @@ static void write_meshIsland(WriteData* wd, MeshIsland* mi)
 	BKE_shard_free(mi->temp, true);
 	mi->temp = NULL;
 
-	writedata(wd, DATA, sizeof(float) * 3 * mi->vertex_count, mi->vertno);
+	writedata(wd, DATA, sizeof(short) * 3 * mi->vertex_count, mi->vertno);
 
 	writestruct(wd, DATA, "RigidBodyOb", 1, mi->rigidbody);
 	writedata(wd, DATA, sizeof(int) * mi->neighbor_count, mi->neighbor_ids);
